backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
supabase>=2.0.0,<3.0.0
python-dotenv>=1.0.0
